refactor into python package
