+++ pyproject.toml
@@ -1,0 +1,28 @@
+[tool.poetry]
+name = "lfelib"
+version = "0.0.1"
+description = "Python routines for working with GPS time series from a variety of processing centers"
+authors = ["Ariane Ducellier <ducela@uw.edu>"]
+license = "GNU General Public License v3.0"
+readme = "README.md"
+homepage = "https://github.com/ArianeDucellier/Incubator2020"
+repository = "https://github.com/ArianeDucellier/Incubator2020"
+keywords = ["seismology", "geophysics", "scientific"]
+
+[tool.poetry.dependencies]
+python = "^3.7"
+numpy = "^1.16"
+scipy = "^1.4"
+matplotlib = "^3.1"
+obspy = "^1.1"
+scikit-learn = "^0.22"
+
+[tool.poetry.dev-dependencies]
+pytest = "^5.2"
+pytest-cov = "^2.8"
+black = {version = "^19.10b0", allow-prereleases = true}
+#dunamai = "^1" #currently failing
+
+[build-system]
+requires = ["poetry>=1.0"]
+build-backend = "poetry.masonry.api"
