+++ tests/test_lfelib.py
@@ -1,0 +1,12 @@
+import lfelib
+
+
+def test_version_string():
+    assert isinstance(lfelib.__version__, str)
+
+# Placeholder tests
+def test_foo():
+    assert True
+
+def test_bar():
+    assert True
